Moved GameHandler to handlers folder
